fix(clippy): clear all -D warnings on workspace
cargo clippy --workspace --all-targets -- -D warnings now exits 0.
Mixed mechanical fixes (cargo clippy --fix applied where it could) plus
hand fixes for the remainder:

  * #[allow(dead_code)] on rng.rs Rng impl (next_word/read_reg/write_reg
    reserved for an interior-mutability rewrite of read_u32; documented).
  * checked_div in cortex_m / riscv / rom_thunks instead of explicit
    'if m == 0 { … } else { n / m }' patterns.
  * Trim doc-list overindentation in system_stub.rs.
  * Replace single-arm match with if in usb_serial_jtag.rs write().
  * Remove unused Bus + DebugControl imports.
  * #[allow(dead_code)] on DAP protocol structs whose fields are
    deserialized for spec compliance but not yet consumed.
  * sort_by_key + std::cmp::Reverse instead of sort_by closure in dap.
  * Drop .map(|a| a) identity in dap.
  * Canonical *self Clone for Copy struct in firmware-l476-demo.
  * # Safety docs on systick/reset/default handlers.
  * #[allow(dead_code)] on TIM1 register-map constants reserved for docs.
  * Remove unnecessary parentheses in xtensa_regs test.

Unblocks PR #81 and PR #82's core-integrity (Run Clippy) step.

+++ crates/core/src/decoder/arm.rs
@@ -1102,7 +1102,7 @@ pub fn decode_thumb_32(h1: u16, h2: u16) -> Instruction {
         let d = (h1 >> 6) & 1;
         let rn = (h1 & 0xF) as u8;
         let vd = ((h2 >> 12) & 0xF) as u8;
-        let imm8 = (h2 & 0xFF) as u16;
+        let imm8 = h2 & 0xFF;
         let sd = (vd << 1) | (d as u8);
         let imm = imm8 << 2;
         let add = u != 0;
@@ -1114,7 +1114,7 @@ pub fn decode_thumb_32(h1: u16, h2: u16) -> Instruction {
         let d = (h1 >> 6) & 1;
         let rn = (h1 & 0xF) as u8;
         let vd = ((h2 >> 12) & 0xF) as u8;
-        let imm8 = (h2 & 0xFF) as u16;
+        let imm8 = h2 & 0xFF;
         let sd = (vd << 1) | (d as u8);
         let imm = imm8 << 2;
         let add = u != 0;
@@ -1348,7 +1348,7 @@ pub fn decode_thumb_32(h1: u16, h2: u16) -> Instruction {
         let imm3 = (h2 >> 12) & 7;
         let rd = ((h2 >> 8) & 0xF) as u8;
         let imm8 = h2 & 0xFF;
-        let imm12 = ((i as u16) << 11) | ((imm3 as u16) << 8) | imm8;
+        let imm12 = (i << 11) | (imm3 << 8) | imm8;
         if rn == 15 {
             return Instruction::Adr { rd, imm: imm12 };
         } else {
@@ -1364,7 +1364,7 @@ pub fn decode_thumb_32(h1: u16, h2: u16) -> Instruction {
         let imm3 = (h2 >> 12) & 7;
         let rd = ((h2 >> 8) & 0xF) as u8;
         let imm8 = h2 & 0xFF;
-        let imm12 = ((i as u16) << 11) | ((imm3 as u16) << 8) | imm8;
+        let imm12 = (i << 11) | (imm3 << 8) | imm8;
         if rn == 15 {
             // ADR.W with negative offset — encoded as PC - imm. The
             // existing Instruction::Adr models PC + imm, so we encode the
